fix(typing): resolve remaining mypy strict typing errors across core domain

+++ backend/app/services/assessment_service.py
@@ -190,7 +190,7 @@ class AssessmentService:
                 pillars.append(
                     FinancialHealthPillarResponse(
                         name=p_name.replace("_", " ").title(),
-                        score=float(s) if s is not None else 0.0,
+                        score=int(s) if s is not None else 0,
                         health_status=h_stat,
                         positive_reason_codes=p_data.get("positive_reason_codes", []),
                         adverse_reason_codes=p_data.get("adverse_reason_codes", []),
